feat(controller): Unification of utils scripts, removal of simultaneous execution of multiple scrapings to execute just one entry.
