--- src/utils/removeMask.js
@@ -1,3 +1,0 @@
-export function removeMask(number) {
-    return number.replace(/[^\d]/g, '');
-}
