--- src/utils/selectors.js
@@ -1,15 +1,0 @@
-export const selectors = {
-    input: '#NI',
-    message: '#mensagem',
-
-    
-    cpfBtn: '#conEdital-2 > div.cel-content > a.btn.btn-primary.btn-one',
-    cpfOkBtn: '.ui-dialog-buttonset > button',
-    cpfFormSelecao: '#FrmSelecao',
-    cpfACertification: '#FrmSelecao > a:nth-child(6)',
-    // cpfACertification: '#FrmSelecao > a:nth-child(3)',
-
-    cpfTableMessage: '#main-container > div > table > tbody > tr > td',
-
-    cnpjBtn: '#conEdital-3 > div.cel-content > a.btn.btn-primary.btn-one',
-};
